--- tools/evm-spec-tester/src/statetest/command.rs
@@ -1,20 +1,0 @@
-use crate::util::make_configuration;
-use cfx_config::Configuration;
-use clap::Args;
-use std::{path::PathBuf, sync::Arc};
-
-/// ethereum statetest doc: https://eest.ethereum.org/main/consuming_tests/state_test/
-#[derive(Args, Debug)]
-pub struct StateTestCmd {
-    /// Paths to state test files or directories
-    #[arg(required = true)]
-    pub(super) paths: Vec<PathBuf>,
-
-    /// Conflux client configuration
-    #[arg(short, long, value_parser = make_configuration, default_value = "", help = "Path to the configuration file")]
-    pub(super) config: Arc<Configuration>,
-
-    /// Only run tests matching this string
-    #[arg(short, long, value_name = "Matches")]
-    pub(super) matches: Option<String>,
-}
