+++ tools/evm-spec-tester/src/statetest/mod.rs
@@ -1,175 +1,66 @@
-pub(crate) mod command;
-pub mod unit_tester;
-mod utils;
+mod unit_tester;
 
-use cfx_executor::machine::{Machine, VmFactory};
-use cfxcore::verification::VerificationConfig;
-use eest_types::StateTestSuite;
-use itertools::Itertools;
-use std::{path::PathBuf, sync::Arc};
+use cfx_config::Configuration;
+use clap::Args;
+use std::{
+    path::{Path, PathBuf},
+    sync::Arc,
+};
 
-use crate::{util::find_all_json_tests, TestError};
-use command::StateTestCmd;
-use unit_tester::UnitTester;
-use utils::skip_test;
+use crate::{
+    test_cmd_runner::EestTestCmdTrait,
+    util::{contains_meta_dir, make_configuration},
+};
+pub use unit_tester::StateUnitTester;
 
-impl StateTestCmd {
-    /// Runs `statetest` command.
-    pub fn run(&self) -> bool {
-        let mut success = true;
-        for path in &self.paths {
-            if !path.exists() {
-                panic!("Path not exists: {:?}", path);
-            }
+/// ethereum statetest doc: https://eest.ethereum.org/main/consuming_tests/state_test/
+#[derive(Args, Debug, Clone)]
+pub struct StateTestCmd {
+    /// Paths to state test files or directories
+    #[arg(required = true)]
+    pub(super) paths: Vec<PathBuf>,
 
-            let test_files = find_all_json_tests(path);
+    /// Conflux client configuration
+    #[arg(short, long, value_parser = make_configuration, default_value = "", help = "Path to the configuration file")]
+    pub(super) config: Arc<Configuration>,
 
-            if test_files.is_empty() {
-                error!("No fixtures found in directory: {:?}", path);
-                continue;
-            }
-
-            match self.run_file_tests(test_files, path) {
-                Ok(true) => {}
-                Ok(false) => {
-                    success = false;
-                }
-                Err(_) => {
-                    success = false;
-                    warn!("Failed to run tests in directory: {:?}", path);
-                    continue;
-                }
-            }
-        }
-        success
-    }
-
-    fn run_file_tests(
-        &self, test_files: Vec<PathBuf>, path: &PathBuf,
-    ) -> Result<bool, String> {
-        let mut success = true;
-        info!(
-            "Running {} TestSuites in {}",
-            test_files.len(),
-            path.display()
-        );
-
-        let machine = {
-            let vm_factory = VmFactory::new(1024 * 32);
-            Arc::new(Machine::new_with_builtin(
-                self.config.common_params(),
-                vm_factory,
-            ))
-        };
-
-        let verification = self.config.verification_config(machine.clone());
-
-        let mut skipped_suite = 0;
-        let mut load_err_suite = 0;
-
-        let mut success_units = 0;
-        let mut skipped_units = 0;
-        let mut total_executions = 0;
-
-        let mut error_list = vec![];
-
-        for path in test_files {
-            if skip_test(&path) {
-                skipped_suite += 1;
-                continue;
-            }
-
-            let (success_cnt, skipped_cnt, transact_cnt, errors) =
-                match SuiteTester::load(&path) {
-                    Ok(tester) => tester.run(
-                        &machine,
-                        &verification,
-                        self.matches.as_deref(),
-                    ),
-                    Err(err_msg) => {
-                        warn!(
-                            "TestSuite load failed. path: {:?}, error: {}",
-                            path, err_msg
-                        );
-                        success = false;
-                        load_err_suite += 1;
-                        continue;
-                    }
-                };
-
-            success_units += success_cnt;
-            skipped_units += skipped_cnt;
-            total_executions += transact_cnt;
-
-            error_list.extend(errors);
-        }
-
-        let error_units = error_list.len();
-
-        for (path, units) in
-            &error_list.into_iter().chunk_by(|err| err.path.clone())
-        {
-            success = false;
-            println!("\nPath {path} fails:");
-            for TestError { name, kind, .. } in units {
-                println!("\t{name}: {kind}");
-            }
-        }
-
-        println!("\n\nSkipped TestSuites: {}", skipped_suite);
-        println!("Load Failed TestSuites: {}", load_err_suite);
-        println!("Success Units: {}", success_units);
-        println!("Skipped Units: {}", skipped_units);
-        println!("Error Units  : {}", error_units);
-        println!("Total Executions: {}", total_executions);
-
-        Ok(success)
-    }
+    /// Only run tests matching this string
+    #[arg(short, long, value_name = "Matches")]
+    pub(super) matches: Option<String>,
 }
 
-struct SuiteTester {
-    path: String,
-    suite: StateTestSuite,
-}
+impl EestTestCmdTrait for StateTestCmd {
+    fn get_matches(&self) -> &Option<String> { &self.matches }
 
-impl SuiteTester {
-    pub fn load(path: &PathBuf) -> Result<Self, String> {
-        let s = std::fs::read_to_string(&path).map_err(|e| e.to_string())?;
-        let suite: StateTestSuite =
-            serde_json::from_str(&s).map_err(|e| e.to_string())?;
+    fn get_paths(&self) -> &Vec<PathBuf> { &self.paths }
 
-        let path = path.to_string_lossy().into_owned();
-        Ok(Self { path, suite })
-    }
+    fn get_config(&self) -> &Arc<Configuration> { &self.config }
 
-    fn run(
-        self, machine: &Machine, verification: &VerificationConfig,
-        matches: Option<&str>,
-    ) -> (usize, usize, usize, Vec<TestError>) {
-        if matches.is_some() {
-            trace!("Running TestUnit: {}", self.path);
-        } else {
-            debug!("Running TestUnit: {}", self.path);
+    fn skip_test(&self, path: &Path) -> bool {
+        if contains_meta_dir(path) {
+            return true;
         }
 
-        let mut error_list = vec![];
-        let mut success_cnt = 0;
-        let mut skipped_cnt = 0;
-        let mut transact_cnt = 0;
-        for (name, unit) in self.suite.0 {
-            let unit_tester = UnitTester::new(&self.path, name, unit);
-            match unit_tester.run(&machine, verification, matches) {
-                Ok(cnt) => {
-                    transact_cnt += cnt;
-                    if cnt > 0 {
-                        success_cnt += 1;
-                    } else {
-                        skipped_cnt += 1;
-                    }
-                }
-                Err(e) => error_list.push(e),
-            }
-        }
-        (success_cnt, skipped_cnt, transact_cnt, error_list)
+        let name = path.file_name().unwrap().to_str().unwrap();
+
+        matches!(
+            name,
+            // Tests not valid at Prague
+            "intrinsicCancun.json"
+
+        // Unreasonable test cases and also skipped by revm (fails in revm)
+        | "RevertInCreateInInitCreate2Paris.json"
+        | "create2collisionStorageParis.json"
+        | "dynamicAccountOverwriteEmpty_Paris.json"
+        | "InitCollisionParis.json"
+        | "RevertInCreateInInit_Paris.json"
+
+        // ## These tests are passing, but they take a lot of time to execute so we are going to skip them.
+        // | "loopExp.json"
+        // | "Call50000_sha256.json"
+        // | "static_Call50000_sha256.json"
+        | "loopMul.json"
+        | "CALLBlake2f_MaxRounds.json"
+        )
     }
 }
